requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Financial Data & Technical Analysis
yfinance
pandas
pandas-ta

# AI/LLM Integration (Multi-Provider Support)
openai==1.54.3
anthropic==0.42.0
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
